Deprecate PropertyCache
