+++ fixture-monkey-api/src/main/java/com/navercorp/fixturemonkey/api/collection/IteratorCache.java
@@ -27,7 +27,7 @@ import org.apiguardian.api.API.Status;
 
 @API(since = "0.4.0", status = Status.MAINTAINED)
 public final class IteratorCache {
-	private static final LruCache<Iterator<?>, List<?>> ITERATOR_TO_LIST = new LruCache<>(2000);
+	private static final LruCache<Iterator<?>, List<?>> ITERATOR_TO_LIST = new LruCache<>(2048);
 
 	public static List<?> getList(Iterator<?> iterator) {
 		if (ITERATOR_TO_LIST.containsKey(iterator)) {
